pyproject.toml
[project]
name = "multimodal-diarization"
version = "0.1.0"
description = "Prototype backend for multimodal speaker diarization ingest"
requires-python = ">=3.10"
dependencies = [
  "fastapi==0.110.0",
  "uvicorn[standard]==0.29.0",
  "aiortc==1.7.0",
  "numpy>=1.24.0",
  "torch>=2.1.0",
  "torchaudio>=2.1.0",
  "python-dotenv>=1.0.1",
  "pyannote.audio>=3.1.1",
  "huggingface_hub>=0.22.0",
  "omegaconf>=2.3.0",
  "webrtcvad>=2.0.10",
  "faster-whisper>=1.0.0",
  "groq>=0.4.0",
  "convex>=0.7.0",
  "elevenlabs>=0.2.0",
  "face-recognition>=1.3.0",
  "opencv-python-headless>=4.11.0.86",
  "sarvamai>=0.1.22",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.uv]
package = false
